Disable deploy to s3
Problem: deploy to chrome web store conflicts with deploying to s3,
because packing .crx with auto-update requires update_url in the
manifest; on the other hand, deployment to chrome web store requires
this field to be absent.

Solution: disable deployment to s3, as deploying the plugin to web
store covers all cases.