Improve tab handling
Problem: the extension icon is only updated when the user is looking
at the page and it does not update thereafter. This is bad as very
often the icon misreports the status of the page.

Solution: update icon for all tabs (even if they are not
active). Also, start fetching article status as soon as url is known.