Properly handle deleted articles
Problem: extension does not detect deletion when article was removed
from the frontend. This leads to multiple additional issues with the
page: it is always reported as saved, impossibe to read or save
again.

Cause: the issue was caused by extension ignoring empty answers,
while in reality they mean there is no article for given url.

Solution: when empty response arrives, delete article for given url.